+++ packages/runtime-base/src/hooks.ts
@@ -3,6 +3,17 @@ import { asArray } from './utils/asArray'
 
 type AnyFunc = (...args: any[]) => any
 
+enum HookInvokeState {
+  pausing = 'pausing',
+  resuming = 'resuming'
+}
+
+interface HookSharedState {
+  state: HookInvokeState
+  stack: [string, Tap, any, any[]][]
+  hooksNameList: MorHookNames[]
+}
+
 interface Tap {
   name: string
   type: 'sync' | 'async'
@@ -16,14 +27,16 @@ interface Tap {
 export class SyncHook {
   name: string
   taps: Tap[]
+  sharedState: HookSharedState
 
   /**
    * @constructor
-   * @param name Hook 名称
+   * @param name - Hook 名称
    */
-  constructor(name) {
+  constructor(name: string, sharedState?: HookSharedState) {
     this.name = name || ''
     this.taps = []
+    this.sharedState = sharedState
   }
 
   /**
@@ -38,7 +51,7 @@ export class SyncHook {
    * @param name Hook 名称
    */
   alias(name: string): SyncHook {
-    const aliasHook = new SyncHook(name)
+    const aliasHook = new SyncHook(name, this.sharedState)
 
     // 这里直接使用 taps 数组, 方便 alias Hook 共用
     aliasHook.taps = this.taps
@@ -93,12 +106,28 @@ export class SyncHook {
     })
 
     for (const tap of taps) {
-      try {
-        tap.fn.call(context, ...args)
-      } catch (err) {
-        logger.error(this.name, tap.name, err)
+      // 暂停中的 hook 将所有调用保存到堆栈中，等待后续恢复
+      if (this.isPausing()) {
+        this.sharedState.stack.push([this.name, tap, context, args])
+      } else {
+        try {
+          tap.fn.call(context, ...args)
+        } catch (err) {
+          logger.error(this.name, tap.name, err)
+        }
       }
     }
+  }
+
+  isPausing() {
+    const state = this.sharedState
+    // 当触发了 $hooks.pause 暂停，若未传入需要指定暂停的 hooks 则暂停所有生命周期触发
+    // 若传入了某些指定的 hooks 数组，则只暂停这些传入 hooks
+    if (state?.state !== HookInvokeState.pausing) return false
+    if (state.hooksNameList?.length === 0) return true
+    if (state.hooksNameList.indexOf(this.name as MorHookNames) !== -1)
+      return true
+    return false
   }
 }
 
@@ -209,9 +238,18 @@ export interface MorHooks {
    * 在 Component 的 onError 生命周期触发
    */
   componentOnError: SyncHook
+  /**
+   * 暂停所有或部分 hooks 的执行并将相关 hooks 的触发保存在堆栈中
+   * @param hooksNameList - 需要暂停的 hook 名称，如未传入则代表暂停所有
+   */
+  pause: (hooksNameList?: MorHookNames[]) => void
+  /**
+   * 恢复 hooks 的执行，并将之前保存在堆栈中的 hooks 按顺序依次触发
+   */
+  resume: () => void
 }
 
-export type MorHookNames = keyof MorHooks
+export type MorHookNames = keyof Omit<MorHooks, 'pause' | 'resume'>
 
 type Reason = string
 
@@ -233,44 +271,75 @@ const HOOKS_INSTANCES: Record<
  * @returns hooks 对象
  */
 export function createHooks(reason: Reason): MorHooks {
-  const appOnConstruct = new SyncHook('appOnConstruct')
-  const pageOnConstructHook = new SyncHook('pageOnConstruct')
-  const componentOnInitHook = new SyncHook('componentOnInit')
-  const componentDidMountHook = new SyncHook('componentDidMount')
-  const componentDidUnmountHook = new SyncHook('componentDidUnmount')
-  const componentOnError = new SyncHook('componentOnError')
+  const sharedState: HookSharedState = {
+    state: HookInvokeState.resuming,
+    stack: [],
+    hooksNameList: []
+  }
+
+  const appOnConstruct = new SyncHook('appOnConstruct', sharedState)
+  const pageOnConstructHook = new SyncHook('pageOnConstruct', sharedState)
+  const componentOnInitHook = new SyncHook('componentOnInit', sharedState)
+  const componentDidMountHook = new SyncHook('componentDidMount', sharedState)
+  const componentDidUnmountHook = new SyncHook(
+    'componentDidUnmount',
+    sharedState
+  )
+  const componentOnError = new SyncHook('componentOnError', sharedState)
 
   const hooks: MorHooks = {
     /* App 相关 hooks */
     appOnConstruct: appOnConstruct,
     // appOnInit 已废弃, 这里出于兼容性暂不移除
     appOnInit: appOnConstruct.alias('appOnInit'),
-    appOnLaunch: new SyncHook('appOnLaunch'),
-    appOnError: new SyncHook('appOnError'),
-    appOnShow: new SyncHook('appOnShow'),
-    appOnHide: new SyncHook('appOnHide'),
-    appOnPageNotFound: new SyncHook('appOnPageNotFound'),
-    appOnUnhandledRejection: new SyncHook('appOnUnhandledRejection'),
+    appOnLaunch: new SyncHook('appOnLaunch', sharedState),
+    appOnError: new SyncHook('appOnError', sharedState),
+    appOnShow: new SyncHook('appOnShow', sharedState),
+    appOnHide: new SyncHook('appOnHide', sharedState),
+    appOnPageNotFound: new SyncHook('appOnPageNotFound', sharedState),
+    appOnUnhandledRejection: new SyncHook(
+      'appOnUnhandledRejection',
+      sharedState
+    ),
 
     /* Page 相关 hooks */
     pageOnConstruct: pageOnConstructHook,
     // pageOnInit 已废弃, 这里出于兼容性暂不移除
     pageOnInit: pageOnConstructHook.alias('pageOnInit'),
-    pageOnLoad: new SyncHook('pageOnLoad'),
-    pageOnReady: new SyncHook('pageOnReady'),
-    pageOnShow: new SyncHook('pageOnShow'),
-    pageOnHide: new SyncHook('pageOnHide'),
-    pageOnUnload: new SyncHook('pageOnUnload'),
+    pageOnLoad: new SyncHook('pageOnLoad', sharedState),
+    pageOnReady: new SyncHook('pageOnReady', sharedState),
+    pageOnShow: new SyncHook('pageOnShow', sharedState),
+    pageOnHide: new SyncHook('pageOnHide', sharedState),
+    pageOnUnload: new SyncHook('pageOnUnload', sharedState),
 
     /* Component 相关 hooks */
-    componentOnConstruct: new SyncHook('componentOnConstruct'),
+    componentOnConstruct: new SyncHook('componentOnConstruct', sharedState),
     componentOnInit: componentOnInitHook,
     componentOnCreated: componentOnInitHook.alias('componentOnCreated'),
     componentDidMount: componentDidMountHook,
     componentOnAttached: componentDidMountHook.alias('componentOnAttached'),
     componentDidUnmount: componentDidUnmountHook,
     componentOnDetached: componentDidUnmountHook.alias('componentOnDetached'),
-    componentOnError: componentOnError
+    componentOnError: componentOnError,
+
+    pause(hooksNameList) {
+      sharedState.state = HookInvokeState.pausing
+      sharedState.hooksNameList = hooksNameList || []
+    },
+
+    resume() {
+      sharedState.state = HookInvokeState.resuming
+      let stackItem = sharedState.stack.shift()
+      while (stackItem) {
+        const [name, tap, context, args] = stackItem
+        try {
+          tap?.fn.call(context, ...args)
+        } catch (error) {
+          logger.error(name, tap.name, error)
+        }
+        stackItem = sharedState.stack.shift()
+      }
+    }
   }
 
   // 记录创建的所有 hooks