+++ packages/utils/src/constants.ts
@@ -57,8 +57,14 @@ export type CompileScriptTargetType = ObjectValues<typeof CompileScriptTarget>
  * - miniprogram: 以小程序的方式编译或集成
  * - plugin: 以插件的方式编译或集成
  * - subpackage: 以插件的方式编译或集成
+ * - component: 以组件的方式编译
  */
-export const CompileTypes = objectEnum(['miniprogram', 'plugin', 'subpackage'])
+export const CompileTypes = objectEnum([
+  'miniprogram',
+  'plugin',
+  'subpackage',
+  'component'
+])
 
 /**
  * Entry 文件类型