+++ packages/plugin-compiler-dingding/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.68](https://github.com/eleme/morjs/compare/v1.0.67...v1.0.68) (2023-07-14)
+
+**Note:** Version bump only for package @morjs/plugin-compiler-dingding
+
+
+
+
+
 ## [1.0.59](https://github.com/eleme/morjs/compare/v1.0.58...v1.0.59) (2023-06-25)
 
 **Note:** Version bump only for package @morjs/plugin-compiler-dingding